Refactor props handling to support custom props
The new functions StatWithProps and ReadDirWithProps - which are wrapped
by their original versions now using default props - allow for
specifying a list of props to request from the server with the PROPFIND
request.

The return File uses the Sys() method of the os.FileInfo interface to
return the underlying props, and a convenience method to retrieve string
props from them.